+++ Hazel/src/Hazel/Rendering/Textures/Texture.h
@@ -7,8 +7,8 @@ namespace Hazel
 	public:
 		virtual ~Texture() = default;
 
-		virtual size_t GetWidth() const = 0;
+		virtual float GetWidth() const = 0;
 		virtual void SetData(const void *data) = 0;
-		virtual void Bind(size_t slot = 0) const = 0;
+		virtual void Bind(uint32_t slot = 0) const = 0;
 	};
 }
