+++ Hazel/src/Hazel/Rendering/Textures/TextureInfo.h
@@ -8,8 +8,8 @@ namespace Hazel
 {
 	struct TextureInfo
 	{
-		size_t Width = 0;
-		size_t Height = 0;
+		float Width = 0;
+		float Height = 0;
 		ColorFormat ColorFormat = ColorFormat::Rgba;
 		TextureFiltering MinFilter = TextureFiltering::Linear;
 		TextureFiltering MagFilter = TextureFiltering::Linear;
